Update tag profile for OCD and OCP
https://issues.redhat.com/browse/ENT-3838
https://issues.redhat.com/browse/ENT-3839

I went ahead and removed the small bit of RHEL config we had from the
file, since it's not in use yet (Let's handle remaining pieces in one
big batch).

I also reworked the tests to check consistency of the tag profile file
rather than looking for explicit values.

Testing
-------
Given this is a just a config change, needs to be tested w/ changes from
https://issues.redhat.com/browse/ENT-3871