Template PromQL account query lookups
Use the QueryBuilder to template the account lookup queries when
pulling the enabled accounts from prometheus.

Account lookup query templates can be defined as a property map:
  rhsm-subscriptions.metering.prometheus.metric.queryTemplates

The key specified in this property map can be specified as a
TagMetric.accountQueryKey in tag_profile.yaml. By default, a key
of 'default' is assumed if not specified.

The template can use any properties of the associated TagMetric class
and query parameters can be defined via the TagMetric.