Suppress certain stacktraces from error messages.
The reason for this is that frequently the location an exception is
created is quite distant from the location where it is logged.  Our
exceptions are generally logged in the BaseExceptionMapper class but the
application class that threw the exception has no way of communicating
that the exception thrown is mundane and does not warrant a stacktrace.
The MDC (mapped diagnostic context) is a thread local data structure
that we can use as a communications channel.

This commit creates an OnMdcEvaluator that suppresses stacktraces from
error log messages based on the content of the MDC.  The intent is for
the developer to drop a known key set to "true" into the MDC at the time
an exception is thrown.  If the OnMdcEvaluator is set to look for that
key, the stacktrace will be suppressed.

This commit suppresses the stacktraces for requests that use an invalid
SLA or usage and for requests specify a granularity too fine for the
given product.