Make various fixes to dependencyJson task
Notably, when called from any given project, the task will now do that
project and any subprojects' dependencies. In other words, called from
the root as `./gradlew -q :dependencyJson`, you'll get output for any
services we have in the repo.

Called from a given subproject, you'll get a report for that specific
component. (e.g. `./gradlew -q rhsm-client:dependencyJson` outputs
only the deps for that subproject/library).

This allows us to use in the short term the same script/tooling as we
have been (with minor tweaks) for reporting out dependencies.

I switched to using the `default` configuration, as this includes both
compile-time and run-time deps. What we had before seemed to be
missing some deps!

Also, I modified the code to walk the dependency tree, as I found out
the methods we were using weren't doing that!

I also put the logic in a `doLast` block so it doesn't execute all the
time. Whoops!