Fix double instance monthly totals calculation
When adding support for instance_hours metering, we changed
the event_source and event_type to be more metric specific.
After it was deployed, when metrics gathering happens
the metrics for a given month would include any older events
with the old static source and type, causing the instance
hours to be doubled.

This patch adds a liquibase changeset that re-aligns the
event table's data with what is now expected in terms of
event_type and event_source.

Because there are environments that already have the
instance_hours functionality running, this patch is a
little cumbersome.

First the unique constraint on the events table has to be
removed because converting old records to the new style
can cause duplicate records once the data updates are run.
We then need to remove the duplicates and add back the
unique constraint.