Extract conduit classes to separate project
https://issues.redhat.com/browse/ENT-3937

- Extract system-conduit README
- Migrate to separate image for swatch-system-conduit deployment
- Migrate rhsm-conduit code to swatch-system-conduit

The following are notable details:

To run conduit, use `./gradlew swatch-system-conduit:bootRun` from the
root of the checkout, or use `../gradlew bootRun` from the
swatch-system-conduit subfolder.

`orgsync` still exists as a profile, and it activates the job to
schedule up syncs for all opted in orgs.

`Dockerfile` builds a separate image for the conduit, but must be build
from the root. (e.g. `podman build . -f
swatch-system-conduit/Dockerfile`).