Read Clowder JSON in via a Spring EnvironmentPostProcessor
This commit takes a Clowder JSON file specified by the
"rhsm-subscriptions.clowder.json-resource-location", reads it into
memory, flattens the JSON structure into Java properties, then pushes
those properties into the Spring configuration environment.

This process occurs early during initialization so that those generated
properties can then be referenced elsewhere.  The primary goal is to
assign the Clowder properties to an intermediate value.  That
intermediate value can then be succinctly overridden or provided in
environments where Clowder is unavailable (e.g. development).

As an example, take the following JSON

    { "kafka": {
        "brokers": [{
          "hostname": "localhost",
        }]
    }}

This commit flattens that JSON to the property
"clowder.kafka.brokers[0].hostname" (the prefix "clowder" is added by
the EnvironmentPostProcessor).  Then in the application.yaml file, we
can assign that property to "KAFKA_BOOTSTRAP_HOST" which is a variable
normally populated from the environment and referenced elsewhere in the
code.

An environment variable or CLI argument defined value for
"KAFKA_BOOTSTRAP_HOST" supersedes a value set in the file, so the value
can be easily overridden.  In the case of local development where
Clowder is unavailable, developers should be able to continue their
standard development process that defined necessary runtime
configuration through environment variables, CLI arguments, or
externally defined configuration files.

Default values should be provided for the intermediate variables;
otherwise, an exception will be thrown during application startup if
the clowder prefixed properties cannot be resolved.