ENT-4319 Add syncOffering(sku) to OfferingSyncController
This adds a method that was missing in the implementation for a
different ticket, ENT-2664. Since this simplifies how to sync an
offering, the callers OfferingJmxBean and OfferingWorker are updated to
use the new method.

Since syncing is only exposed by that one method, the logging around
syncing is updated to make it easier to determine if an offering has
been synced or not synced and why.