ENT-3376 Enable subscription sync for cron job
This will set SUBSCRIPTION_SYNC_ENABLED to true for the
subscription-sync cron job. With the updates to app-interface
(https://gitlab.cee.redhat.com/service/app-interface/-/merge_requests/25597)
which sets it to false for QA, STAGE, and PROD, this allows only CI to
run the periodic subscription sync, until we're ready.