ENT-4371 Sync EMPTY Usage and ServiceLevel not null
Offerings should only sync (that is, be saved to the db) when the
fetched offering differs from the already synced offering. However,
offerings were always being synced.

The cause turned out to be that when the fetched offering is assembled
in UpstreamProductData, if there are no Usage or ServiceLevel fields
set, then they would be null. However, when the same offering is loaded
from the DB, those null fields would instead be the EMPTY enum value
instead.

The fix is to make UpstreamProductData never make Usage or ServiceLevel
be null, but instead be EMPTY.