Skip hourly processing if there are no events
https://issues.redhat.com/browse/ENT-4372

Otherwise, we pull in whole accounts which is wasteful (and causes
problematic amounts of load).

This is a forward-port of #590 to the `release` branch.

Also note that given this cleanly merges to `develop` as well, so once
we perform the R.C. process, it'll get into develop naturally. (Though
I'm happy to do a third PR instead if anyone wants).