Read Clowder configuration using JSON Path
For certain configuration keys, the Clowder JSON provides a list instead
of a map.  Additionally, the list can be in an indeterminate order.  For
example,

```
"endpoints": [
  {
    "app": "rhsm-clowdapp",
    "hostname": "rhsm-clowdapp-api.rhsm.svc",
    "name": "api",
    "port": 8000
  },
  {
    "app": "rhsm-clowdapp",
    "hostname": "rhsm-clowdapp-metrics-worker.rhsm.svc",
    "name": "metrics-worker",
    "port": 8000
  },
  {
    "app": "rbac",
    "hostname": "rbac-service.rhsm.svc",
    "name": "service",
    "port": 8000
  }
]
```

The previous method of locating Clowder configuration values simply
flattened the JSON, allowing us to use an expression like
`endpoints[0].hostname`.  But, if the order of the list is variable, we
can't specify a static index and expect it to work.

This commit takes properties prefixed with `clowder.` and evaluates them
as JSON path expressions.  This approach allows us to us expressions
like `endpoints[?(@.app == 'rbac')]` which filters the list to find
items that have the `app` key equal to `rbac`.

See https://goessner.net/articles/JsonPath/ for the JSON Path syntax.