Tally hourly snapshots by Service Type
Determine a set of configured service types and do a
seperate tally for each of them for a specific account.

NOTES:
* Removes ProductProfile from the MetricUsageCollector and
  instead collects metrics based on Service Type.
* Needed to enable snapshot generation for RHOSAK in the API spec.
* Filter events by service type when pulling from the DB.

ATTENTION:
Because Spring Data does not support querying JSONB, we needed
to query via a native query, and also are currently unable to
write a test around the repository methods because HSQLDB dialect
does not understand the JSONB queries.