Fix InventoryAccountUsageCollectorTest random failures
Specifically, removesDuplicateHostRecords was failing because of a
few bad assumptions.

First, the test assumed that a specific instance would be removed, but
the iteration is on a HashMap, so iteration is dependent on the key
hash, so not particular predictable. (see invocation of
handleDuplicateHost from collect in InventoryAccountUsageCollector).

Second, because the method updates hosts by populating information from
HBI, it changes the instanceId that the test set to its inventory ID.

This resulted in two entries to the Map in AccountServiceInventory, one
with the instanceId set in the unit test, and the other with the
inventoryId as key.

I altered the logic to not overwrite the instanceId if it's already been
set.

Also, given which Host is removed is abritrary, I removed the assertion
that was about the "dupe" key being removed.

Testing
-------

I was able to repro consistently by changing
InventoryHostFactTestHelper#createBaseHost to set a fixed UUID:

```
baseFacts.setInventoryId(UUID.fromString("54f765c8-9814-4eb7-b1f2-ec116c768057"));
```

Run

```
./gradlew :test --tests "org.candlepin.subscriptions.tally.InventoryAccountUsageCollectorTest.removesDuplicateHostRecords"
```