Set management server port to 8000
Clowder requires use of port 8000.  The docs state "Public services in a
ClowdApp MUST be running on port 8000. If your app uses a different
port, such as 8080 for its API, the app must be changed to expose port
8000."

Actuator isn't a public service, but I have a hunch that we need it on
port 8000 to be able to access in a useful way.  If I'm wrong, we can
easily revert this single change.