Look up subscription ID based on offering name
Removed the ProductProfileRegistry from the MarketplaceSubscriptionIdProvider
so that product config can all be done in the tag profile. Also updated the
offering lookup query to be based on Offering name instead of role.

Added rhosak tag to eligible-swatch-product-ids in the marketplace config so that
it will be included in metrics that are sent to marketplace.

Updated the product name mappings in the tag profile config to reflect the values
that will come from the product service to be ready for turning on subscription
table functionality. Also added a RHOSAK mapping.

Updated the offering product names in DB for OCP and OSD, as well as added RHOSAK