Fix RHSM api regex
Quite simply, the regex we had in our code was from the
rhsm-api/pinhead, and was buggy. The regex has already been fixed in
rhsm-api, so all we need to do is fix on our side.

A failing example is `2021-10-27T20:17:23Z`, which fails the current
regex, because `20` doesn't pass `(0[0-9]|1[0-9]|2[1-4])`. (only 00-19
and 21-24 pass the buggy regex).

Tip: you can you "context actions" (typically alt+enter) to check the
regex directly in intellij, which can be used for verification.