Correct trap on hawtio-proxy script
Previously we were trapping `INT TERM EXIT`.  EXIT is actually a
pseudo-signal that is bash specific.  Bash runs EXIT when the script
exits under any circumstances.  The end result is that when the user
hits ^C, the INT signal runs the cleanup trap and then the EXIT signal
runs it again.  Running the trap twice meant that the second time
through, the cleanup function was trying to kill a process that no
longer existed.