Quiet down a log message
If the connection to prometheus isn't made (because you don't have the
token-refresher running for example) this message is printed on every
retry.  The repeated output makes it appear as though something is
happening when it isn't.  It would be nice to have the log output
something like "Retry 2 of 4..." or similar but the Spring retry
template is several methods up in the stack and adding that message
would be a messy prospect.