Skip sending summnary with no measurements
When producing tally summaries, skip any that do not have
measurements associated with them.

This was done to prevent marketplace API errors when sending
an event with 0 measurements. This can happen when event metrics
are updated via the JMX endpoint and a metric event was no longer
found during the refresh.