Handle truncated IP addresses when gathering NIC info for HBI
Candlepin has a limit of 255 chars for a given fact. When a fact is sent
that has a longer value, the value is truncated, and the last 3
characters are replaced with "...".

In ENT-4260, we began to populate network interfaces. However, the code
did not filter out truncated IP addresses. This commit corrects that
issue and also refactors some of the IP address handling code to better
isolate logic in methods.