Correct some annotations on our Hawtio customization
https://issues.redhat.com/browse/ENT-4658

When running Hawtio on a different management port, the order of bean
initialization changed such that our HawtioContextListener was being
instantiated before the HawtioManagementConfiguration.  Since the former
depended on a bean in the latter, this caused a start up failure.

This commit changes the annotations on the HawtioContextListener so that
it will load under the same circumstances the
HawtioManagementConfiguration bean loads and the former is marked as
having a dependency on the latter.

Additionally, when management services (Spring Actuator) are run on a
different port from the main application, the management services exist
in a child context descended from the main context.  So that our
configuration will exist in that same context, we annotate it with
ManagementContextConfiguration and put a special entry in the
META-INF/spring.factories file.

See the ManagementContextAutoConfiguration class for details.