Allow access to the Spring Actuator root context
Hitting the root context of Spring Actuator displays all the actuators
that are available over HTTP.  It's a useful shortcut and before this
patch, hitting that URL would require authentication.  Even when
deployed in DEV mode, the security framework still requires the
x-rh-identity header to be present and have a valid structure.