ENT-4662: set connection TTL for HttpClient instances
Per
https://hc.apache.org/httpcomponents-client-4.5.x/current/tutorial/html/connmgmt.html
when the server does not send an explicit value for Keep-Alive, the
connection manager will assume the connection can be kept open
indefinitely. Marketplace does not send a Keep-Alive. However, in
testing, QE confirmed that connections are terminated after ~505 seconds
Therefore, we set a connection TTL of 5 minutes, well below the critical
threshold.