Rename marketplace-worker for clarity
https://issues.redhat.com/browse/ENT-4711

In order to be consistent with the WIP in #746, I renamed the deployment
to swatch-billing-provider-rh-marketplace. In many other places I
renamed to rhmarketplace or rh-marketplace or rhMarketplace depending on
the context. This will also require some changes in the deployment, so
there will be some app-interface MRs to go along with this.