Query the Cloudigrade user API for user existence before further calls
Querying the Cloudigrade concurrent API with for a user that does not
exist results in a non-200 return code.  We interpret this is a failure
and retry the call and log an error.  This commit queries the internal
Cloudigrade user API to check for the existence of a user prior to
making any subsequent calls and therefore avoids entering error handling
code.