Use Spring's Resource abstraction to handle loading keystores
Previously we had the keystore properties bound to a File field.  I was
seeing failures when the rhsm-conduit.rhsm.keystore property was set to
empty string which could happen during unit tests that run without a
full configuration.

Using the Spring Resource abstraction avoids this issue since Spring
isn't trying to immediately create a File object.  Instead it creates a
Resource object and we later check to make sure the Resource is valid
and readable.  As a bonus, Spring will handle opening the resource as a
stream for us so we can remove some boiler plate methods in
HttpClientProperties that were doing that.