Add github workflows to manage stage branch alias
Testing
-------

Apply to a fork of this repo's `main` branch.

Create a branch named `stage`. (This action updates the branch, but
won't create it.

Try pushing to the `main` branch of your fork (be careful not to push
to upstream `main`).

Also try opening a PR against your fork's `main` branch, and merge it.
Observe that for this action also, the action is triggered.