Fix permissions when podman_run.sh falls back to docker
We got into a state where the CI user can't cleanup the workspace
because file permissions are off. This is addressed by having the
effective UID of the container process be the same as the calling user.
I found a consistent way to do this for both docker and podman is by
using --user=$UID along with setting the USERNS setting for podman to
`keep-id`, which can conveniently be done with an env var.