Remove Repeatable annotation from validator annotations
This code was incorrect.  It claimed to be for the purpose of enabling a
validator in a type declaration (e.g. List<@IpAddress String>) but the
Repeatable annotation and the associated declaration actually just
enabled the annotation to be applied multiple times to the same field,
parameter, etc.  This isn't a very useful property for these annotations
so I'm removing it and the incorrect comment.