Remove external api spec download during build.
Update the build to remove the download of an external API spec during
compilation. This speeds up the build and prevents the need to download
from an external repository for a source file during our compilation.
The upside is that if all the dependencies have been downloaded we
no longer make network calls during a regular build. In addition it
allows us to remove an external gradle plugin from our configuration.

This requires us to be more intentional about when we update clients
for the one of 3 updates that was pointing to the latest from a
repository instead of a particular git hash. It does bring consistency
so that now all of the API clients that we interact with are being built
in house.