Update deployment configuration for PSK authentication
Previously, the PSKs were stored in a JSON blob assigned to one
configuration property.  The blob was then parsed at runtime to get the
PSKs for individual applications.  Spring Boot allows us to bind
configuration properties to a map so I simplified the code to use this
functionality instead of parsing JSON.

This patch makes the deployment changes necessitated by the conversion
to using configurations properties bound to a map.