Adjust usage latency to meet AWS SLA
Usage metering now occurs every hour on the half hour
and is delayed by only 30 minutes relative to when the job
is run. This means that if the metering job runs at 2:30PM,
metrics will be gathered for 1pm (the 1pm - 2pm range).

When the metrics query is run, it is run against the
2pm - 2:59:59.99999pm range since the query does a 1h look back
and the data actually represents usage for 1pm.

The hourly tally schedule has been updated to run with a
1h delay, relative to when the job is run. This give the
metrics tasks 30m to complete and to have events populated.