Remove unnecessary Mockito matcher calls
Mockito argument matchers aren't necessary when all arguments are meant
to be matched with "eq".

To quote from Sonar's description "the default matching behavior (i.e.
without argument matchers) uses equals(). If only the matcher
org.mockito.ArgumentMatchers.eq() is used, the call is equivalent to the
call without matchers, i.e. the eq() is not necessary and can be
omitted. The resulting code is shorter and easier to read."

This is Sonar error code java:S6068