Realign and simplify range handling w/ metrics
During verification of #953, I noticed that the way the metrics
ranges are handled was a bit inconsistent. Specifically, when we
persist metrics, we treat the data point timestamp as the "end" of
the range and then calculate back to the beginning of the range by
subtracting an hour. For a data point that represents 12pm-1pm, we
record `timestamp` as `12pm` and `expiration` as `1pm`, but the
handling of the active account query and task interpretation of
ranges were different.

I also noticed that the JMX operation `performCustomMetering` was not
functioning properly when given a range longer than 1 hour, which I
addressed.

These changes make the entry points to metrics gathering also follow
this convention - the start of the range and end of the range for
12pm to 1pm are consistently represented across the data flow.

Testing
-------
Follow instructions in [Generating Prometheus Metric Data](https://docs.engineering.redhat.com/display/ENT/Generating+Prometheus+Metric+Data)

Run the metering job with debug logging to see the range queried:

```
LOGGING_LEVEL_ORG_CANDLEPIN_SUBSCRIPTIONS_METERING=DEBUG \
    PROM_URL="http://localhost:9090/api/v1/" \
    SPRING_PROFILES_ACTIVE=metering-job,kafka-queue \
    ./gradlew :bootRun
```

Note the log messages reflect that the time range of interest is the previous hour:

```
2022-05-05 14:02:09,958 [thread=restartedMain] [DEBUG] [org.candlepin.subscriptions.metering.service.prometheus.PrometheusService] - Fetching metrics from prometheus: 2022-05-05T14:00Z -> 2022-05-05T14:00Z [Step: 3600]
...
2022-05-05 14:02:09,965 [thread=restartedMain] [DEBUG] [org.candlepin.subscriptions.metering.service.prometheus.PrometheusAccountSource] - Querying for active accounts for range [2022-05-05T13:00Z, 2022-05-05T14:00Z)
...
2022-05-05 14:02:09,966 [thread=restartedMain] [INFO ] [org.candlepin.subscriptions.task.queue.kafka.KafkaTaskQueue] - Queuing task: TaskDescriptor[groupId: platform.rhsm-subscriptions.metering-tasks, taskType: METRICS_COLLECTION, args: [metric: [Transfer-gibibytes], productTag: [rhosak], start: [2022-05-05T13:00Z], end: [2022-05-05T14:00Z], account: [account123]]]
```

Note 14:00 in the above example is the appropriate timestamp
given our `[1h]` range vector look back.

Run the metrics worker to pick up the queued tasks:

```
LOGGING_LEVEL_ORG_CANDLEPIN_SUBSCRIPTIONS_METERING=DEBUG \
    SUBSCRIPTION_USE_STUB=true \
    USER_USE_STUB=true \
    DEV_MODE=true \
    PROM_URL="http://localhost:9090/api/v1/" \
    ./gradlew :bootRun
```

Observe that start and end time reflect *only* the previous hour,
and that the DB query lines up with the time range (the `[1h]`
range vector is important here too):

```
2022-05-05 14:03:58,611 [thread=metering-task-processor-0-C-1] [INFO ] [org.candlepin.subscriptions.metering.service.prometheus.PrometheusService] - Fetching metrics from prometheus: 2022-05-05T14:00Z -> 2022-05-05T14:00Z [Step: 3600]
2022-05-05 14:03:58,611 [thread=metering-task-processor-0-C-1] [DEBUG] [org.candlepin.subscriptions.metering.service.prometheus.PrometheusService] - Running prometheus range query: Start: 1651759200 End: 1651759200 Step: 3600, Query: kafka_id:kafka_broker_quota_totalstorageusedbytes:max_over_time1h_gibibytes%20*%20on(_id)%20group_right%20min_over_time(subscription_labels%7Bproduct=%22rhosak%22,%20ebs_account=%22account123%22,%20billing_model=%22marketplace%22,%20support=~%22Premium%7CStandard%7CSelf-Support%7CNone%22%7D%5B1h%5D)
2022-05-05 14:03:58,616 [thread=metering-task-processor-0-C-1] [DEBUG] [org.candlepin.subscriptions.metering.service.prometheus.PrometheusMeteringController] - Looking for events in range [2022-05-05T13:00Z, 2022-05-05T14:00Z)
```

Verify the behavior of the MeterJmxBean changes.
Specifically, note I added some input validation on rangeInMinutes.

Running `performCustomMetering` with 60 minutes queries for active
accounts across a 1 hour range, and 120 minutes looks for active
accounts across a two-hour range.

Running `performCustomMeteringForAccount` directly queues tasks
with the proper ranges.

Entering something not on the hourly boundary for `rangeInMinutes`
(e.g. `61`) causes an exception.