Fix issue with unit tests failing on Jenkins
Two tests in the InternalTallyResourceTest were failing in Jenkins. They
ran fine locally and I have been unable to duplicate the Jenkins issue
with the podman_run.sh script.  The functionality these failing tests
exercise, the security constraints on the URLs under the
"**/internal/**" path, is also exercised in
InternalSubscriptionResourceTest.  So instead of going on a wild goose
chase, I'm going to cut my loses and delete the failing (and redundant)
tests.