Add endpoints to force kafka seeks in swatch-producer-aws
Testing
-------

Add the following to BillableUsageProcessor.process:

```java
    try {
      Thread.sleep(10000);
    } catch (InterruptedException e) {
      throw new RuntimeException(e);
    }
```

Run via:

```
QUARKUS_LOG_LEVEL=DEBUG ./gradlew swatch-producer-aws:quarkusDev
```

Enqueue some messages, and try the endpoints at http://localhost:8000/api/swatch-producer-aws/swagger-ui/

Try seeking forward and backwards.

Note that quarkus buffers messages on the topic, and doesn't expose a
way to clear the buffers. However, it does detect when this has
happened, and will log a `DEBUG` message such as:

```
2022-06-07 13:28:00,521 DEBUG [io.sma.rea.mes.kafka] (vert.x-eventloop-thread-3) SRMSG18239: Received acknowledgement for record 2 on 'platform.rhsm-subscriptions.billable-usage-0' (consumer group: 'swatch-producer-aws'). Ignoring it because the partition is not assigned to the consume anymore. Record will likely be processed again. Current assignments are [platform.rhsm-subscriptions.billable-usage-0].
```

In practice, the behavior is that we may process (redundantly) messages
in the buffer before the seek takes effect. Note that in spite of
redunantly processing these messages, the committed offset remains
seeked. (Once the buffer is cleared, we begin to process messages from
the seeked offset).