Set appropriate tags for our internal APIs
The OpenApi generator uses the tag field for the class name of the
generated Api class.  Having two tags for one operation means that two
classes get generated and both contain that operation.  Possibly is is
desirable in some cases, but not in ours.

This patch removes the multiple tags from the
internal-subscriptions-sync-api operations and provides a better name
for the internal-tally-api operation.