Add endpoint to sync remittance
The endpoint will fetch all of the latest snapshots for each
metric for an account, filter them based on the configured
billing window and billing validity, and run them through the
remittance process.

If remittance already exists for a metric, a new remittance
value will not be calculated as the metric has already been
billed.

When remittance is synced, the new remittance value is calculated
and no billable usage will be sent as we consider this usage already
billed for.

To test:
1. Load events.
2. Do tally and capture remittance.
3. Delete remittance.
4. Hit new API and verify remittance is equal.