Send TallySummary messages after snapshots are persisted
Remittance logic requires that applicable snapshots be persisted before the
resulting TallySummary message is consumed. When processing a lot of events that result in
a lot of snapshots (for example during a re-tally), message consumption can sometimes be a little
of snapshot persistance. This can result in an incorrect calculated remittance.

Moved the message sending outside of the DB transaction. When remittance
is updated, we require that all relevant snapshots be persisted so that
our totals are accurate since TallySummary messages can be received out
of order.