Adjust logging levels with Splunk HEC appender
Prior to this patch, the root logger was set to INFO level. This setting
printed the INFO events for our code, but also for third-party code.
This patch changes the settings so that only loggers in the
org.candlepin namespace emit INFO messages.  All other loggers are set
to WARN level.