Add Kafka SASL configuration for Spring Kafka and Clowder
The SASL configuration is contained in a separate YAML document.
Attempting to reconcile development deployment and Clowder deployment
all within one configuration document proved to be too challenging.
Specifically, getting the unit test runs to ignore the SASL
configuration (since the in-memory broker wasn't configured for SASL)
and getting the in-line PEM for the certificate to parse correctly as a
default value for a property.

Instead, the TLS/SASL configuration is in a separate YAML document that
is only loaded when Spring detects that it is running under
Kubernetes/Openshift.  The heuristic it uses for making this
determination is the existence of some environment variables.  Right
now, this works for us, but there is no facility for customizing the
heuristic should Openshift drift from Kubernetes in some fashion.