Tolerate missing account number for opt-in checks
When opt-in is checked, only the OrgConfig related to
principal is checked (opt-in considered complete).

The ReportingAccessService also needed to be updated to
account for a missing account number to allow the requests
to succeed. Since the reportingEnabled flag is associated
with the account config only, if the account was not specified
during the request, we allow reporting if the org was opted in.