Update grafana dashboard with latest schema
This patch is the current dashboard just re-exported on today's date to
reflect version and schema changes.

Also update documentation with some notes about creating the initial
export.  Basically, selecting the export that is for "sharing
externally" seems to add some hardcoded values that are undesirable.
I'm not 100% certain that selecting for external sharing is always wrong
though so I recommended developers try both exports and select the one
with the least invasive changes.