Add default Kafka connection properties
Without defaults, clowder-quarkus-config-source won't set the properties
at all.  And if the properties are not set and we are in a Kafka
environment that communicates over TLS, Kafka has an odd bug where
attempting to connect to a TLS enabled server without having a security
protocol set (e.g. clowder-quarkus-config-source didn't set it), the
client throws an OutOfMemoryException.  See
https://issues.apache.org/jira/browse/KAFKA-4090

Kevin Howell figured out this issue and likely saved my sanity in the
process.