Move gradle build out of dockerfiles.
This converges all docker buildds to the pattern
use by swatch-producer-aws.

The overall goal of this PR is to reduce the number of
general gradle builds that are done when the
pr_check.sh script is run.

Add a .dockerignore that requires specific enablement
into each docker context root used in the builds.

Use a blanket block in the .dockerignore to force
specific folders to be enabled.