Always load Kafka JAAS post-processor
The KafkaJaasBeanPostProcessor always needs to be loaded.
Theoretically, we could limit it to load only when Kafka is in use.  In
practice, however, I've run into multiple situations where Kafka is
being used but the JAAS post processor isn't being created.  Without the
post-processor, pod deployment in the ephemeral environments immediately
fails because Kafka doesn't have correct JAAS configurations.

Enabling the bean when it doesn't need to be enabled incurs an extremely
minimal cost and given the difficulties of untangling the web of
profiles we have, I think it will just make everyone's life easier if
the JAAS post-processor is always created.