Enable kafka-queue profile for subscription-sync
The kafka-queue profile needs to be present for Spring to load the Kafka
related configuration in swatch-core/application.yaml.  That file has a
YAML document with a spring.config.activate.on-profile directive that
requires the `kafka-queue` profile to be present in order to load.
Without that configuration, any Kafka communication that needs to be
configured from Clowder provided values won't work.