Add internal API for hourly tally
An internal hourly tally API endpoint was added to run the hourly tally
operation for a single account.

If the application is run with the RHSM_SUBSCRIPTIONS_ENABLE_SYNCHRONOUS_OPERATIONS
env var set, the API will support synchronous operations. The
hourly tally can be run synchronously if the following request header
is present:

x-rh-swatch-synchronous-request: true