Correct error with how SWatch was checking for users in Cloudigrade
In a recent change, the API spec for Cloudigrade's internal API was
modified to change a query parameter into a header parameter.  This
change had the effect of making the REST call to Cloudigrade "/users"
instead of "/users?username=$NAME".  The end result was that our calls
to Cloudigrade to see if a user existed instead returned the entire list
of users.  Compounding the problem, the code in CloudigradeService was
not written carefully.  That user existence check code only anticipated
0 or 1 results.  When the results were an entire list of users, the code
interpreted that as "yes, the user in question exists."  When
CloudigradeAccountUsageCollector got the answer "yes, the user exists",
it would then try to query Cloudigrade for the user's concurrency info
(authenticating as that user).  That user does not exist in Cloudigrade
and an auth error results.

This patch also dispenses with the attempt to authenticate to
Cloudigrade using the X-RH-CLOUDIGRADE-ACCOUNT-NUMBER header.  For calls
to the internal Cloudigrade API, only the PSK is necessary.  For calls
to the external concurrency API, we were attempting to set an
authentication header for any information we had.  That could mean
setting X-RH-CLOUDIGRADE-ACCOUNT-NUMBER or X-RH-CLOUDIGRADE-ORG-ID or
both.  Rather than deal with all the combinations, I've altered the code
to use Org ID only since that's the way the world is moving.