Add MDC properties for OrgId and Account id when processing BillableUsage messages
This is an experimental add to set org_id and account_id when processing
BillableUsage events in swatch-producer-aws. This will create a properties
block on the messages in splunk hec. After verification of how well this works
we can examine whether it is worth spreading the pattern more widely or
reverting.