Refactor Spring Data Specification usage to be less abstract
Rather than using a generic framework for building query predicates,
this patch uses the Spring Data Specification API to build just the
predicates we need.  We can compose those predicates together with the
`and` method.  The result is code that is easier to read and follow and
extensible if and when we need to add additional conditions to the
query.