Refactor hypervisor report category to be an either/or
Trying to treat the HypervisorReportCategory choice as a set just leads
to very awkward code.  ReportCategory values can only map to one
HypervisorReportCategory so the set would only ever have one value if a
ReportCategory is specified.  If no ReportCategory was given, the
HypervisorReportCategory would have all of its values (HYPERVISOR and
NON_HYPERVISOR) but the specification code would just ignore that since
there was no reason to apply a filter when the filter would match
everything.