Add stricter validation to IP addresses in facts
Also modify some of the logic around how we handle validation for lists
of items versus singleton items.  We want to check lists of items for
truncation ("127.0.0.1,192...") and log if the value is truncated.  For
singleton items, we don't expect any commas to split on or truncation
tokens.  If there happens to be a comma or truncation token, the value
will fail the ordinary validation anyway.

See SWATCH-684