Capture time properly in splunk HEC events
This applies specifically to our spring-boot based deployments.

Since we were using a custom event header serializer, we need to supply
`time` appropriately.

Testing
=======

Check the ephemeral deployment for this PR. Check in splunk for
`namespace=${ephemeral_namespace} host=swatch-tally-service*`. Note that
the timestamp emitted in `message` and recorded for the splunk event
should match down to the microsecond.