Properly set account during hourly tally
We were setting account_number on account calculations
based on AccountInventoryService.account which may not
exist if there is no account associated with an org.

This patch will set it based on the first
host instance that has a value set. This way
the account number is in essance coming from
the event that created the host.