Add swatch.java-conventions to swatch-core-test
When I switched my JAVA_HOME to java 17, I got some odd errors like

```
> Could not resolve all task dependencies for configuration ':testRuntimeClasspath'.
   > Could not resolve project :swatch-core-test.
     Required by:
         project :
      > No matching variant of project :swatch-core-test was found. The consumer was configured to find a runtime of a library compatible with Java 11, packaged as a jar, preferably optimized for standard JVMs, and its dependencies declared externally but
...
```

which was happening because the swatch-core-test was being built w/
default compatibility (17 when JAVA_HOME points to Java 17). By applying
the swatch.java-conventions, its explicitly set to Java 11, along with
the rest of the swatch modules.