Use Guava's Cartisian product method to calculate UsageCalculation.Keys
The product list being passed in to InventoryAccountUsageCollector is
actually declared as a Set in TallySnapshotController.  The method
signature for collect was up-casting it to a Collection for no real
reason.  If we keep it as a set and make sets of the SLA and Usage, we
can use the convenient Cartesian product functionality to eliminate an
ugly bit of nested loops.